Added the tableParser to the main system
Incorporated the tableParser Rafey made to the main system. It runs before the processing of the other elements (before the classifier). It returns the enitre xml file with processed tables included, so in the final showing of the elements on frontend we must find a different way to display each table.