Correct pagenr and elementnr
The program should now be better at setting correct page and element numbers. It first checks if an element has a label tag, if not then it uses regex to check if the element text contains the elementnr, and if that also fails it defaults to using the nr which grobid gave it. But this number from grobid could be wrong.